+++ src/Symfony/Component/Security/CHANGELOG.md
@@ -7,6 +7,8 @@ CHANGELOG
  * Added access decision strategy to override access decisions by voter service priority
  * Added `IS_ANONYMOUS`, `IS_REMEMBERED`, `IS_IMPERSONATOR`
  * Hash the persistent RememberMe token value in database.
+ * Added `LogoutEvent` to allow custom logout listeners.
+ * Deprecated `LogoutSuccessHandlerInterface` and `LogoutHandlerInterface` in favor of listening on the `LogoutEvent`.
 
 5.0.0
 -----